+++ Changelog.md
@@ -1,5 +1,6 @@
-# Unreleased
+# 1.0.3 - 2015-12-03
 
+- Dependency update to support require support for the `links` element.
 - When a parser returns an error it is sometimes useful to inspect the parse result. This now gets passed back to the handler function and can be used to print more information, such as parser annotations, when an error occurs.
 
 # 1.0.2 - 2015-11-30